Added 4 buttons to the main menu activity
Added 4 buttons to the main menu activity, also added a gridlayout which
is the parent of the buttons. Changed the id of each button to what it
should be and the text of each button. Don't have our good sketches on
me right now but will change these appropriately when I have the
sketches in front of me.