Added Text View and Changed Colors
Added a text view to the main menu as an intro message. Also edited the
colors in the colors.xml file for the app, feel free to change, just
wanted to put something that fit the park theme in. Also added a string
resource for the text on the main menu.