Added Weather To MainMenu
Filled in the empty top space in the main menu with some weather info.
The temperature is a borderless button. I had trouble with the
toolbar/actionbar in main menu for god knows what reason so I had to
increase the top margins on the weather stuff so it wouldn't overlap
with the bar. Probably has something to do with the fact I chose a basic
activity thing to start it, wrong choice apparently.